Updated name and type of callstack region id

+++ src/model/src/datamodel/rocprofvis_dm_stack_record.h
@@ -19,7 +19,7 @@ class StackFrameRecord
         // @param args - stack frame method arguments
         // @param line - stack frame code line
         // @param depth - stack frame depth
-        StackFrameRecord(rocprofvis_dm_charptr_t symbol, rocprofvis_dm_charptr_t args, rocprofvis_dm_charptr_t line, uint32_t depth, uint32_t id):
+        StackFrameRecord(rocprofvis_dm_charptr_t symbol, rocprofvis_dm_charptr_t args, rocprofvis_dm_charptr_t line, uint32_t depth, uint64_t id):
             m_symbol(symbol), m_args(args), m_code_line(line), m_depth(depth), m_id(id) {};
         // Returns pointer to symbol string
         rocprofvis_dm_charptr_t         Symbol() {return m_symbol.c_str();}
@@ -30,7 +30,7 @@ class StackFrameRecord
         // returns stack frame depth
         uint32_t                        Depth() {return m_depth;}
         // returns region id
-        uint32_t                        RegionId() {return m_id;}
+        uint64_t                        RegionId() {return m_id;}
     private:
         // stack frame symbol string
         rocprofvis_dm_string_t           m_symbol;
@@ -41,7 +41,7 @@ class StackFrameRecord
         // stack frame depth
         uint32_t                         m_depth;
         // region id
-        uint32_t                         m_id;
+        uint64_t                         m_id;
 };
 
 }  // namespace DataModel
